Grunt Work 1 Added folders and images
Added folders to all sections "audio" and "images" need to migrate
images in section 1 and 2 to their respective folders. Added images and
directories to sections3-8. Lot of images added in section6 "term11" is
missing could not retrieve image.  In section 8 "term68" and "term70"
are missing due they not being images rather texted formatted.